pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fairprs-clin"
version = "0.1.0"
description = "ClinGen PRS-RS–aligned ancestry-stratified PRS evaluation framework"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "Minahal Aisha"},
]
dependencies = [
  "numpy>=1.22",
  "pandas>=1.5",
  "matplotlib>=3.6",
  "pyyaml>=6.0",
]

[project.scripts]
fairprs-clin = "fairprs_clin.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
